server/requirements.txt
eventlet==0.26.1
gevent_socketio==0.3.6
